Add tests covering diagnostics amidst multiple errors
* Refactor diagnostics to avoid maintaining a custom error type
* Mostly centralize diagnostics into Diagnostics.swift
* New tests to verify diagnostics in complex cases with "layered" errors
