+++ Sources/MemberwiseInitMacros/Macros/Support/Diagnostics.swift
@@ -1,105 +1,164 @@
 import SwiftDiagnostics
 import SwiftSyntax
+import SwiftSyntaxMacroExpansion
 
-enum MemberwiseInitMacroDiagnostic: Error, DiagnosticMessage {
-  case defaultAppliedToMultipleBindings
-  case labelAppliedToMultipleBindings
-  case labelConflictsWithProperty(String)
-  case labelConflictsWithAnotherLabel(String)
-  case invalidDeclarationKind(DeclGroupSyntax)
-  case invalidSwiftLabel
-  case missingTypeForVarProperty
-  case tupleDestructuringInProperty
+// MARK: - Diagnose VariableDeclSyntax
 
-  private var rawValue: String {
-    switch self {
-    case .defaultAppliedToMultipleBindings:
-      ".defaultAppliedToMultipleBindings"
+func diagnoseMultipleConfigurations(variable: VariableDeclSyntax) -> [Diagnostic]? {
+  guard variable.customConfigurationAttributes.count > 1 else { return nil }
 
-    case .labelAppliedToMultipleBindings:
-      ".labelAppliedToMultipleBindings"
-
-    case .invalidDeclarationKind(let declGroup):
-      ".invalidDeclarationKind(\(declGroup.kind))"
-
-    case .invalidSwiftLabel:
-      ".invalidLabel"
-
-    case .labelConflictsWithProperty(let label):
-      ".labelCollidesWithProperty(\(label))"
-
-    case .labelConflictsWithAnotherLabel(let label):
-      ".labelCollidesWithAnotherLabel(\(label))"
-
-    case .missingTypeForVarProperty:
-      ".missingTypeForVarProperty"
-
-    case .tupleDestructuringInProperty:
-      ".tupleUsedInProperty"
-    }
-  }
-
-  var severity: DiagnosticSeverity { .error }
-
-  var message: String {
-    switch self {
-    case .defaultAppliedToMultipleBindings:
-      return "Custom 'default' can't be applied to multiple bindings"
-
-    case .labelAppliedToMultipleBindings:
-      return """
-        Custom 'label' can't be applied to multiple bindings
-        """
-
-    case let .invalidDeclarationKind(declGroup):
-      return """
-        @MemberwiseInit can only be attached to a struct, class, or actor; \
-        not to \(declGroup.descriptiveDeclKind(withArticle: true)).
-        """
-
-    case .invalidSwiftLabel:
-      return "Invalid label value"
-
-    case let .labelConflictsWithProperty(label):
-      return "Label '\(label)' conflicts with a property name"
-
-    case let .labelConflictsWithAnotherLabel(label):
-      return "Label '\(label)' conflicts with another label"
-
-    case .missingTypeForVarProperty:
-      return
-        "@MemberwiseInit requires a type annotation."
-
-    case .tupleDestructuringInProperty:
-      return """
-        @MemberwiseInit does not support tuple destructuring for property declarations. \
-        Use multiple declarations instead.
-        """
-    }
-  }
-
-  var diagnosticID: MessageID {
-    .init(domain: "MemberwiseInitMacro", id: rawValue)
+  return variable.customConfigurationAttributes.dropFirst().map { attribute in
+    Diagnostic(
+      node: attribute,
+      message: MacroExpansionErrorMessage(
+        "Multiple @Init configurations are not supported by @MemberwiseInit"
+      )
+    )
   }
 }
 
-extension DeclGroupSyntax {
-  func descriptiveDeclKind(withArticle article: Bool = false) -> String {
-    switch self {
-    case is ActorDeclSyntax:
-      return article ? "an actor" : "actor"
-    case is ClassDeclSyntax:
-      return article ? "a class" : "class"
-    case is ExtensionDeclSyntax:
-      return article ? "an extension" : "extension"
-    case is ProtocolDeclSyntax:
-      return article ? "a protocol" : "protocol"
-    case is StructDeclSyntax:
-      return article ? "a struct" : "struct"
-    case is EnumDeclSyntax:
-      return article ? "an enum" : "enum"
-    default:
-      return "`\(self.kind)`"
+func diagnoseVariableDecl(
+  customSettings: VariableCustomSettings?,
+  variable: VariableDeclSyntax,
+  targetAccessLevel: AccessLevelModifier
+) -> [Diagnostic] {
+  let customSettingsDiagnostics =
+    customSettings.map { settings in
+      [
+        diagnoseVariableLabel(customSettings: settings, variable: variable),
+        diagnoseDefaultValueAppliedToMultipleBindings(
+          customSettings: settings,
+          variable: variable
+        ),
+      ].compactMap { $0 }
+    } ?? []
+
+  let accessibilityDiagnostics = [
+    diagnoseAccessibilityLeak(
+      customSettings: customSettings,
+      variable: variable,
+      targetAccessLevel: targetAccessLevel
+    )
+  ].compactMap { $0 }
+
+  return customSettingsDiagnostics + accessibilityDiagnostics
+}
+
+private func diagnoseVariableLabel(
+  customSettings: VariableCustomSettings,
+  variable: VariableDeclSyntax
+) -> Diagnostic? {
+  if let label = customSettings.label,
+    label != "_",
+    variable.bindings.count > 1
+  {
+    return customSettings.diagnosticOnLabel(
+      MacroExpansionErrorMessage("Custom 'label' can't be applied to multiple bindings")
+    )
+  }
+
+  if customSettings.label?.isInvalidSwiftLabel ?? false {
+    return customSettings.diagnosticOnLabelValue(MacroExpansionErrorMessage("Invalid label value"))
+  }
+
+  return nil
+}
+
+private func diagnoseDefaultValueAppliedToMultipleBindings(
+  customSettings: VariableCustomSettings,
+  variable: VariableDeclSyntax
+) -> Diagnostic? {
+  guard
+    customSettings.defaultValue != nil,
+    variable.bindings.count > 1
+  else { return nil }
+
+  return customSettings.diagnosticOnDefault(
+    MacroExpansionErrorMessage("Custom 'default' can't be applied to multiple bindings")
+  )
+}
+
+private func diagnoseAccessibilityLeak(
+  customSettings: VariableCustomSettings?,
+  variable: VariableDeclSyntax,
+  targetAccessLevel: AccessLevelModifier
+) -> Diagnostic? {
+  let effectiveAccessLevel = customSettings?.accessLevel ?? variable.accessLevel
+
+  guard
+    targetAccessLevel > effectiveAccessLevel,
+    !variable.isFullyInitializedLet
+  else { return nil }
+
+  let customAccess = variable.customConfigurationArguments?
+    .first?
+    .expression
+    .as(MemberAccessExprSyntax.self)
+
+  let targetNode =
+    customAccess?._syntaxNode
+    ?? (variable.modifiers.isEmpty ? variable._syntaxNode : variable.modifiers._syntaxNode)
+
+  return Diagnostic(
+    node: targetNode,
+    message: MacroExpansionErrorMessage(
+      """
+      @MemberwiseInit(.\(targetAccessLevel)) would leak access to '\(effectiveAccessLevel)' property
+      """
+    )
+  )
+}
+
+// MARK: - Diagnose [PropertyBinding] and [MemberProperty]
+
+func customInitLabelDiagnosticsFor(bindings: [PropertyBinding]) -> [Diagnostic] {
+  var diagnostics: [Diagnostic] = []
+
+  let customLabeledBindings = bindings.filter {
+    $0.variable.customSettings?.label != nil
+  }
+
+  // Diagnose custom label conflicts with another custom label
+  var seenCustomLabels: Set<String> = []
+  for binding in customLabeledBindings {
+    guard
+      let customSettings = binding.variable.customSettings,
+      let label = customSettings.label,
+      label != "_"
+    else { continue }
+    defer { seenCustomLabels.insert(label) }
+    if seenCustomLabels.contains(label) {
+      diagnostics.append(
+        customSettings.diagnosticOnLabelValue(
+          MacroExpansionErrorMessage("Label '\(label)' conflicts with another label")
+        )
+      )
     }
   }
+
+  return diagnostics
+}
+
+func customInitLabelDiagnosticsFor(properties: [MemberProperty]) -> [Diagnostic] {
+  var diagnostics: [Diagnostic] = []
+
+  let propertiesByName = Dictionary(uniqueKeysWithValues: properties.map { ($0.name, $0) })
+
+  // Diagnose custom label conflicts with a property
+  for property in properties {
+    guard
+      let propertyCustomSettings = property.customSettings,
+      let label = propertyCustomSettings.label,
+      let duplicated = propertiesByName[label],
+      duplicated != property
+    else { continue }
+
+    diagnostics.append(
+      propertyCustomSettings.diagnosticOnLabelValue(
+        MacroExpansionErrorMessage("Label '\(label)' conflicts with a property name")
+      )
+    )
+  }
+
+  return diagnostics
 }
